Ajout de #define à la place de valeurs numériques des capteurs de proxy

+++ miniprojet_SlopeFollower/regulation.h
@@ -11,7 +11,7 @@
 // régulateur PID
 int16_t regulator(int16_t angle, int16_t angle_consigne, bool reset);
 
-int16_t esquive(int8_t alert_number);
+int32_t esquive(int8_t alert_number);
 
 //démarrage du thread de régulation
 void regulator_start(void);
